@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 2.5.2+11
+version: 3.0.1+20
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -11,40 +11,48 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_neumorphic:
-    path: ./neomorphic
-  cupertino_icons: ^0.1.3
-  dio: ^3.0.10
-  equatable: ^1.2.3
-  firebase_analytics: ^5.0.2
-  firebase_auth: ^0.18.1+2
-  firebase_core: ^0.5.0+1
-  firebase_database: ^4.1.1
-  fl_chart: ^0.12.0
-  flutter_bloc: ^4.0.1
-  flutter_svg: ^0.19.0
-  flutter_webview_plugin: ^0.3.11
-  html: ^0.14.0+4
-  intl: ^0.16.1
-  lottie: ^0.6.0
-  mantras: ^0.0.3
-  provider: ^4.3.2
-  shared_preferences: ^0.5.12
-  shimmer: ^1.1.2
-  url_launcher: ^5.7.10
-  webview_flutter: ^1.0.1
-  firebase_messaging: ^7.0.3
+  flutter_neumorphic: ^3.1.0
+  adaptive_theme: ^2.2.0
   circular_profile_avatar: ^2.0.0
+  cupertino_icons: ^1.0.3
+  dio: ^4.0.0
+  equatable: ^2.0.2
+  # firebase_analytics: ^8.1.0
+  polls: ^1.0.1
+  introduction_screen: ^2.1.0
+  firebase_auth: ^1.2.0
+  firebase_core: ^1.2.0
+  firebase_database: ^7.0.0
+  firebase_messaging: ^10.0.0
+  fl_chart: ^0.36.1
+  flutter_bloc: ^4.0.1
   flutter_share: ^2.0.0
-  
+  flutter_svg: ^0.22.0
+  flutter_webview_plugin: ^0.4.0
+  html: ^0.15.0
+  intl: ^0.17.0
+  lottie: ^1.0.1
+  mantras: ^0.0.3
+  provider: ^4.0.5
+  share: ^2.0.1
+  shared_preferences: ^2.0.5
+  shimmer: ^2.0.0
+  syncfusion_flutter_pdfviewer: 
+  url_launcher: ^6.0.4
+  string_validator: ^0.3.0
+  image_picker: ^0.8.0+1
+  webview_flutter: ^2.0.7
+  firebase_storage: ^8.1.1
+  giphy_picker: ^2.0.0
+  pull_to_refresh: ^2.0.0
+  profanity_filter: ^1.0.4
+  sweetsheet: ^0.3.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_native_splash: 
-  flutter_launcher_icons:
-
-
+  flutter_launcher_icons: null
+  flutter_native_splash: null
 
 flutter:
   uses-material-design: true
@@ -64,4 +72,3 @@ flutter:
 flutter_native_splash:
   image: assets/images/logo.png
   color: "#FFFFFF"
-
