@@ -1,12 +1,13 @@
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:html/parser.dart';
 
-import '../const.dart';
 import '../local/shared_pref.dart';
 import '../model/Cred.dart';
 import '../model/GeneralResponse.dart';
 import '../model/TokensModel.dart';
 import '../model/User.dart';
+import '../model/hangout/hangUser.dart';
 import '../service/webClientService.dart';
 
 class Auth extends ChangeNotifier {
@@ -18,15 +19,15 @@ class Auth extends ChangeNotifier {
 
   Cred _cred;
   // ignore: unnecessary_getters_setters
-  get user => _user;
+  User get user => _user;
   // ignore: unnecessary_getters_setters
   set user(User u) => _user = u;
   // ignore: unnecessary_getters_setters
-  get cred => _cred;
+  Cred get cred => _cred;
   // ignore: unnecessary_getters_setters
   set cred(Cred u) => _cred = u;
 
-  get token => _tokens;
+  Tokens get token => _tokens;
 
   //  Handle Login
   //  Step 1: Get Cookies
@@ -64,8 +65,8 @@ class Auth extends ChangeNotifier {
       // Submit login Request
       GeneralResponse postResponse =
           await _webClient.postLogin(formData, _tokens.cookies);
-      if (postResponse.status &&
-          postResponse.data.request.uri.toString() == kWebsiteURL + kHomeURL) {
+      print(postResponse.data);
+      if (postResponse.status) {
         //  Login Success
         return postResponse;
       } else {
