@@ -1,5 +1,12 @@
+import 'package:adaptive_theme/adaptive_theme.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+
 import 'routes.dart';
 
-void main() {
-  new Routes();
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  final savedThemeMode = await AdaptiveTheme.getThemeMode();
+  new Routes(savedThemeMode);
 }
